+++ client/css/_fonts.scss
@@ -3,57 +3,69 @@
 
 $open-sans: 'Open Sans', sans-serif;
 
-.fa-microphone-slash { color: $icon-silent; }
-.fa-user { color: $icon-teaching; }
-.fa-users { color: $icon-collab; }
-.fa-laptop { color: $icon-collab; }
+// Icon colors for hangout types
+.fa-microphone-slash { color: $icon-silent; } // red
+.fa-user { color: $icon-teaching; }          // coral
+.fa-users { color: $icon-collab; }           // green
+.fa-laptop { color: $icon-collab; }          // green
 
-a { color: $dodgerblue; }
+// Links
+a { 
+  color: $cta-blue; // coral accent
+  &:hover { color: darken($cta-blue, 10%); cursor: pointer; }
+}
 
+// Hangout headers
 h4.hangout-date,
 h5.in-progress,
 h5.completed {
   display: inline;
   font-size: 1em;
-  letter-spacing: 1;
-  color: $black;
+  letter-spacing: 1px;
+  color: $white;
   font-weight: 500;
   line-height: 1.2;
 }
-h4.completed {
-  color: $warm-grey;
-}
+
+h4.completed { color: $warm-grey; }
+
 h4.hangout-date {
   display: block;
   margin: 2px;
 }
+
 .upcoming {
-    font-style: italic;
-    font-weight: 400;
+  font-style: italic;
+  font-weight: 400;
+  color: $secondary-grey;
 }
 
 h5.in-progress,
 h5.upcoming {
-    color: $primary-grey;
+  color: $primary-grey;
 }
 
 h5.in-progress {
   font-weight: bold;
-  color: #777;
+  color: $secondary-grey;
 }
 
+// Oval indicator
 .oval {
-	@include border-radius(5px);
-	width: 10px;
-	height: 10px;
-	background-color: $danger-red;
-	display: inline-block;
+  @include border-radius(5px);
+  width: 10px;
+  height: 10px;
+  background-color: $hangout-silent; // red
+  display: inline-block;
 }
 
 span.cb-delete:hover {
   cursor: pointer;
+  color: $cta-red; // hover accent
 }
 
+// Search message text
 .search-message {
-    font-weight: 200;
+  font-weight: 200;
+  color: $primary-grey;
 }
