+++ client/css/_inbox.scss
@@ -1,106 +1,111 @@
 #conversations_list {
-	overflow-y: scroll;
-	max-height: 400px;
-	margin-bottom: 20px;
+  overflow-y: scroll;
+  max-height: 400px;
+  margin-bottom: 20px;
 }
+
 .conversation-unread {
   background: $light-grey;
 }
 
 textarea#message-body {
-	resize: vertical;
+  resize: vertical;
 }
 
 #participants {
-	padding: 10px 0px 15px 0px;
+  padding: 10px 0 15px 0;
 }
 
-// conversations
+// ---------------------------
+// Chat bubbles
+// ---------------------------
 .conversation-bubble {
-	color: $white;
-	background: $cta-blue;
-	padding: 10px;
-	@include border-radius(15px); 
-	&:hover {
-		color: $white;
-	}
+  color: $white;
+  background: $cta-blue;
+  padding: 10px;
+  @include border-radius(15px);
+
+  &:hover {
+    color: $white;
+  }
 }
 
 i.submit {
-	display: block;
-	float: right;
-	position: relative;
-	top: -22px;
-	right: 5px;
-	&:hover {
-		color: $cta-blue;
-		cursor: pointer;
-	}
+  display: block;
+  float: right;
+  position: relative;
+  top: -22px;
+  right: 5px;
+
+  &:hover {
+    color: $cta-blue;
+    cursor: pointer;
+  }
 }
 
-// chat bubble
-.ib-clear {clear: both}
+// clearfix
+.ib-clear {
+  clear: both;
+}
+
+// bubble container
 .ib {
-	p { margin: 0;}
+  p {
+    margin: 0;
+  }
 }
+
+// From Me
 .ib-from-me {
-	position:relative;
-	padding:15px;
+  position: relative;
+  padding: 15px;
   margin-bottom: 4px;
-	color: black;
+  color: $black;
+  background: $inbox-green;
   border: 1px solid $light-grey;
-	@include border-radius(25px); 
-	float: right;
+  border-radius: 0.4em;
+  float: right;
+
+  &:after {
+    content: '';
+    position: absolute;
+    right: 0;
+    top: 50%;
+    width: 0;
+    height: 0;
+    border: 20px solid transparent;
+    border-left-color: $inbox-green;
+    border-right: 0;
+    border-bottom: 0;
+    margin-top: -10px;
+    margin-right: -20px;
+  }
 }
+
+// From Them
 .ib-from-them {
-	position:relative;
-	padding:10px 20px;
+  position: relative;
+  padding: 10px 20px;
   margin-bottom: 4px;
+  color: $black;
+  background: $light-grey;
   border: 1px solid $light-grey;
-	@include border-radius(25px);
-	color: black;
+  border-radius: 0.4em;
   float: left;
-	cursor: pointer;
-}
-.ib-from-me {
-	position: relative;
-	right: 20px;
-	background: $inbox-green;
-	border-radius: .4em;
-}
+  cursor: pointer;
 
-.ib-from-me:after {
-	content: '';
-	position: absolute;
-	right: 0;
-	top: 50%;
-	width: 0;
-	height: 0;
-	border: 20px solid transparent;
-	border-left-color: $inbox-green;
-	border-right: 0;
-	border-bottom: 0;
-	margin-top: -10px;
-	margin-right: -20px;
-}
-.ib-from-them {
-	position: relative;
-	left: 20px;
-	background: $light-grey;
-	border-radius: .4em;
-}
-
-.ib-from-them:after {
-	content: '';
-	position: absolute;
-	left: 0;
-	top: 50%;
-	width: 0;
-	height: 0;
-	border: 20px solid transparent;
-	border-right-color: $light-grey;
-	border-left: 0;
-	border-bottom: 0;
-	margin-top: -10px;
-	margin-left: -20px;
+  &:after {
+    content: '';
+    position: absolute;
+    left: 0;
+    top: 50%;
+    width: 0;
+    height: 0;
+    border: 20px solid transparent;
+    border-right-color: $light-grey;
+    border-left: 0;
+    border-bottom: 0;
+    margin-top: -10px;
+    margin-left: -20px;
+  }
 }
