+++ client/css/_slack.scss
@@ -10,25 +10,64 @@
   }
 
   .slack-top {
-    background: linear-gradient(to top, rgba(30, 144, 255, .5) 0%, rgba(30, 144, 255, .5) 0%, rgba(30, 144, 255, .5) 0%, rgba(30, 144, 255, .55) 18%, rgba(30, 144, 255, .64) 46%, rgba(3, 87, 232, .8) 100%), url(/images/language-logos.svg) center repeat;
+    background: linear-gradient(
+        to top,
+        rgba(30, 144, 255, 0.5) 0%,
+        rgba(30, 144, 255, 0.55) 18%,
+        rgba(30, 144, 255, 0.64) 46%,
+        rgba(3, 87, 232, 0.8) 100%
+      ),
+      url(/images/language-logos.svg) center repeat;
+    background-size: cover, auto;
     padding: 100px 0;
-    h2 { @include font-face(700, 32px, bold, $bg-white); }
+
+    h2 {
+      @include font-face(700, 32px, bold, $bg-white);
+      text-align: center;
+    }
   }
 
   .slack-middle {
     background: $bg-white;
     padding: 50px 0;
+
+    p, li {
+      @include font-face(400, 16px, normal, $primary-grey);
+      line-height: 1.7;
     }
+  }
 
   .etiquette {
-    h1 { margin-top: 60px; margin-right: 0px; margin-bottom: 20px; margin-left: 0px;
-      font-size: 34px; font-weight: 700; letter-spacing: -.03rem; color: #333;}
-    h2 { font-size: 27px; font-weight: 600; line-height: 33px; letter-spacing: -1px; margin-right: 0px;
-      margin-bottom: 0px; margin-left: 0px;}
-    h3 { font-size: 15px; font-weight: 800; line-height: 22px; min-height: 22px;
-      text-transform: uppercase; letter-spacing: 1px; margin-top: 0px; margin-right: 0px;
-      margin-bottom: 0px; margin-left: 0px;}
-    hr {margin-top: 1.9em}
+    h1 {
+      margin: 60px 0 20px 0;
+      font-size: 34px;
+      font-weight: 700;
+      letter-spacing: -.03rem;
+      color: $primary-grey;
+    }
+    h2 {
+      font-size: 27px;
+      font-weight: 600;
+      line-height: 33px;
+      letter-spacing: -1px;
+      margin: 0;
+      color: $cta-blue;
+    }
+    h3 {
+      font-size: 15px;
+      font-weight: 800;
+      line-height: 22px;
+      min-height: 22px;
+      text-transform: uppercase;
+      letter-spacing: 1px;
+      margin: 0;
+      color: $primary-grey;
+    }
+    hr {
+      margin-top: 1.9em;
+      border-color: $secondary-grey;
+      opacity: 0.2;
+    }
   }
 
   .slack-hangout {
@@ -40,7 +79,7 @@
     border-top: 1px solid $secondary-grey;
     width: 60px;
     margin-bottom: 1.65em;
-    opacity: .2;
+    opacity: 0.2;
   }
 
   .slack-bottom {
@@ -53,12 +92,13 @@
   }
 
   a.user-popover {
-    .popover { margin-top: -10px;}
+    .popover { margin-top: -10px; }
+
     .popover:hover {
       cursor: default;
     }
   }
 
-  h3.popover-title { margin-top: 0;}
-  h3 { font-weight: 200;}
+  h3.popover-title { margin-top: 0; }
+  h3 { font-weight: 200; color: $primary-grey; }
 }
