+++ client/css/_study_groups.scss
@@ -1,90 +1,131 @@
-  #study-groups {
+#study-groups {
   h2 {
-    margin: 10px 0px;
+    margin: 10px 0;
     @include font-face(600, 24px, normal, $black);
   }
+
   h3 {
     text-transform: uppercase;
     margin: 20px 0 10px 0;
     @include font-face(700, 14px, bold, $black);
   }
+
   .tab-pane h3 {
-    margin: 0px 0px 10px 0px;
+    margin: 0 0 10px 0;
   }
+
   .study-groups-top {
-    background: linear-gradient(to top, rgba(30, 144, 255, .5) 0%, rgba(30, 144, 255, .5) 0%, rgba(30, 144, 255, .5) 0%, rgba(30, 144, 255, .55) 18%, rgba(30, 144, 255, .64) 46%, rgba(3, 87, 232, .8) 100%), url(/images/language-logos.svg) center repeat;
+    background: linear-gradient(
+      to top,
+      rgba(30, 144, 255, 0.5) 0%,
+      rgba(3, 87, 232, 0.8) 100%
+    ), url(/images/language-logos.svg) center repeat;
     padding: 100px 0;
-    h2 { @include font-face(700, 32px, bold, $bg-white); }
-    p { color: $white; }
-    span.label { padding: .3em .6em;}
+
+    h2 {
+      @include font-face(700, 32px, bold, $bg-white);
+    }
+
+    p {
+      color: $white;
+    }
+
+    span.label {
+      padding: 0.3em 0.6em;
+      background: rgba(255, 255, 255, 0.15);
+      border-radius: 3px;
+    }
   }
+
   #newHangout {
     margin-top: 10px;
   }
-  .page-header { margin-top: 0px; }
+
+  .page-header {
+    margin-top: 0;
+  }
+
   .panel-body {
     background: $well;
   }
+
   .panel-default {
-  	border: none;
+    border: none;
   }
+
   .img-circle {
-      width: 50px;
+    width: 50px;
   }
+
   #cancelAddResource {
-  	margin-right: 5px;
+    margin-right: 5px;
   }
-  .well, .jumbotron {
-  	margin: 0px 0px 10px 0px;
+
+  .well,
+  .jumbotron {
+    margin: 0 0 10px 0;
   }
+
   .warning-well {
     border: 1px solid $danger-red;
     color: $danger-red;
   }
+
   .card {
     @media screen and (min-width: 992px) {
       height: 190px;
       padding: 19px 19px 10px;
-      h4 {
-        a {
-          display: block;
-          white-space: nowrap;
-          overflow: hidden;
-          text-overflow: ellipsis;
-        }
-        small.tagline {
-          overflow-y: auto;
-          display: block;
-          max-height: 2.5rem;
-          line-height: 1.5;
-        }
+
+      h4 a {
+        display: block;
+        white-space: nowrap;
+        overflow: hidden;
+        text-overflow: ellipsis;
+      }
+
+      small.tagline {
+        overflow-y: auto;
+        display: block;
+        max-height: 2.5rem;
+        line-height: 1.5;
       }
     }
-    @include breakpoint(large) { height: 190px; padding: 19px 19px 10px; }
+    @include breakpoint(large) {
+      height: 190px;
+      padding: 19px 19px 10px;
+    }
   }
+
   p {
-  	margin: 10px 0px;
+    margin: 10px 0;
   }
+
   p.activity {
-    padding: 0px 15px;
+    padding: 0 15px;
   }
+
   p.overflowtext {
-    margin: 0px;
+    margin: 0;
   }
+
   #info .panel-body {
-  	padding: 10px 0px;
+    padding: 10px 0;
   }
+
   ul#members {
     list-style: none;
-    padding: 0px 15px 10px 15px;
+    padding: 0 15px 10px 15px;
+
     strong {
       color: $black;
       padding-left: 5px;
     }
   }
+
   small.members {
     white-space: nowrap;
   }
+
   a.memberDetail {
     position: relative;
     top: 8px;
@@ -97,11 +138,13 @@
   .label-box {
     margin-top: $margin-base;
   }
+
   #hide-sidebar {
-      @include font-face(700, 22px, normal, $cta-blue);
-      position: relative;
-      top: -10px;
+    @include font-face(700, 22px, normal, $cta-blue);
+    position: relative;
+    top: -10px;
   }
+
   #show-sidebar {
     display: none;
     position: fixed;
@@ -109,16 +152,14 @@
     left: -64px;
     top: 20%;
     background: $cta-blue;
+
     a {
       @include font-face(700, 22px, normal, $white);
       padding: 5px 10px;
       text-decoration: none;
     }
+
     @include border-radius(3px);
-    -webkit-transform: rotate(90deg);
-    -moz-transform: rotate(90deg);
-    -o-transform: rotate(90deg);
-    -ms-transform: rotate(90deg);
     transform: rotate(90deg);
   }
 
@@ -131,48 +172,67 @@
     display: block;
     margin: 10px 0;
     height: 30px;
+    border: none;
+    transition: background-color 0.3s ease;
+
+    &:hover {
+      background-color: darken($cta-blue, 7%);
+    }
   }
 
   .btn-leave-study-group,
   .btn-leave-study-group-1 {
-      background-color: $danger-red;
+    background-color: $danger-red;
+
+    &:hover {
+      background-color: darken($danger-red, 7%);
+    }
   }
 }
 
 .newStudyGroup {
-	fieldset {
-		border: none;
-	}
+  fieldset {
+    border: none;
+  }
 }
 
 .study-groups-list {
   ul {
     list-style: none;
   }
-  ul, li a:link, li a:visited {
+
+  ul,
+  li a:link,
+  li a:visited {
     color: $white;
   }
+
   li {
     margin-bottom: 10px;
   }
+
   h5 {
     padding-left: 40px;
     color: $white;
-    margin: 0px;
+    margin: 0;
     text-transform: uppercase;
   }
-  h5, li.dropdown {
+
+  h5,
+  li.dropdown {
     display: inline;
     text-decoration: none;
   }
-  li.dropdown ul.dropdown-menu {
-    a:link, a:visited {
-      color: $cta-blue;
-      text-decoration: none;
-    }
+
+  li.dropdown ul.dropdown-menu a:link,
+  li.dropdown ul.dropdown-menu a:visited {
+    color: $cta-blue;
+    text-decoration: none;
   }
+
   a.study-group-item {
     display: block;
+
     &:hover {
       color: $hangout-collab;
     }
