+++ client/css/_update_status.scss
@@ -1,20 +1,23 @@
-.silent-tooltip + .tooltip > .tooltip-inner {
-  background-color: $bg-lightgrey;
-  color: $black;
-  text-align: left;
-  border: 1px solid $hangout-silent;
-}
-
-.teach-tooltip + .tooltip > .tooltip-inner {
-  background-color: $bg-lightgrey;
-  color: $black;
-  text-align: left;
-  border: 1px solid $hangout-teaching;
-}
-
+.silent-tooltip + .tooltip > .tooltip-inner,
+.teach-tooltip + .tooltip > .tooltip-inner,
 .collab-tooltip + .tooltip > .tooltip-inner {
   background-color: $bg-lightgrey;
   color: $black;
   text-align: left;
+  font-size: 14px;
+  line-height: 1.4;
+  padding: 8px 12px;
+  border-radius: 4px;
+}
+
+.silent-tooltip + .tooltip > .tooltip-inner {
+  border: 1px solid $hangout-silent;
+}
+
+.teach-tooltip + .tooltip > .tooltip-inner {
+  border: 1px solid $hangout-teaching;
+}
+
+.collab-tooltip + .tooltip > .tooltip-inner {
   border: 1px solid $hangout-collab;
 }
