--- client/templates/hangout/hangout-edit.html
@@ -1,97 +1,0 @@
-<template name="editHangoutModal">
-  <div id="edit-hangout-modal" class="modal fade">
-    <div class="modal-dialog">
-      <div class="modal-content">
-        <div class="modal-header">
-          <h4 class="modal-title text-center">{{_ "edit_a_hangout"}}</h4>
-          <button type="button" class="btn btn-circle btn-outline btn-info pull-right" data-dismiss="modal"><i class="fas fa-times"></i></button>
-        </div>
-        <div class="modal-body">
-          <div class="row">
-            <div class="col-md-7">
-              <div id="date-group" class="row">
-                <div class="col-md-6">
-                  <div class="input-group date" id="start-date-time-picker">
-                    <span class="input-group-addon"><i class="glyphicon glyphicon-time"></i></span>
-                    <input id="start-date-time" class="form-control" type="text" placeholder="Start Time" readonly>
-                  </div>
-                </div>
-                <div class="col-md-6">
-                  <div id="set-end-date-time-block">
-                    <select class="form-control" id="end-date-time">
-                        <option value="25">25 minutes</option>
-                        <option value="60">60 minutes</option>
-                        <option value="90">90 minutes</option>
-                        <option value="120">2 hours</option>
-                        <option value="180">3 hours</option>
-                        <option value="1440">End manually</option>
-                    </select>
-                  </div>
-                </div>
-              </div>
-              <br>
-              <div class="row">
-                <div class="col-md-12">
-                  <input id="topic" class="form-control" type="text" placeholder="Study Session Topic">
-                </div>
-              </div>
-              <br>
-
-              <div class="row">
-                <div class="col-md-12">
-                  <input id="group" class="form-control" type="text" disabled="true">
-                </div>
-              </div>
-              <br>
-
-              <div class="row">
-                <div class="col-md-12">
-                  <div data-editor-host></div>
-                </div>
-              </div>
-            </div>
-            <div id="hangout-options" class="col-md-5 text-left">
-              <div class="btn-group btn-group-vertical" data-toggle="buttons">
-                <label class="btn active">
-                  <input type="radio" name='hangout-type' value="silent" checked><i class="fas fa-circle fa-2x"></i><i class="fas fa-dot-circle fa-2x"></i> <span>  <i class="fas fa-microphone-slash text-danger-color type"></i><span class="title">{{_ "silent_hangout"}}</span></span>
-                </label>
-                <label class="btn">
-                  <input type="radio" name='hangout-type' value="teaching"><i class="fas fa-circle fa-2x"></i><i class="fas fa-dot-circle fa-2x"></i><span> <i class="fas fa-user text-warning-color type"></i><span class="title">{{_ "teaching_hangout"}}</span></span>
-                </label>
-                <label class="btn">
-                  <input type="radio" name='hangout-type' value="collaboration"><i class="fas fa-circle fa-2x"></i><i class="fas fa-dot-circle fa-2x"></i><span> <i class="fas fa-users text-success-color type"></i><span class="title">{{_ "collaboration_hangout"}}</span></span>
-                </label>
-              </div>
-              <div id="external">
-               <label><input type="checkbox" id="external-checkbox" name="externalCheckbox"> Use external platform to host hangout</label>
-               <div id="external-fields" style="display: none;">
-                   <p><label>Join us on:</label>
-                       <input type="text" name="externalButtonText" id="externalButtonText" placeholder="e.g. Zoom, Twitch, etc." class="form-control">
-                   </p>
-                   <p>
-                       <label>URL</label>
-                       <input type="text" name="externalURL" id="externalURL" placeholder="Zoom, Twitch, Google Hangout etc. URL" class="form-control">
-                   </p>
-                   
-              </div>
-              </div>
-            </div>
-            
-          </div>
-          <br>
-          <div class="row">
-            <div class="col-md-12 text-center">
-              <button id="edit-hangout" class="btn btn-cb2">{{_ "save_changes"}}</button>
-            </div>
-          </div>
-          <br>
-          <div class="row">
-            <div class="col-md-12 text-center">
-              <span class="message">{{_ "edit_hangout_hint"}}</span>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</template>
