--- client/templates/hangout/hangout-frame.html
@@ -1,22 +1,0 @@
-<template name="hangoutFrame">
-
-<div id="hangout-button-wrapper">
-  <p class="warning chrome-firefox-warning">Warning: Please load this page using Chrome <img src="/images/chrome.ico" width="25"/> or Firefox <img src="/images/firefox.ico" width="25"/> to avoid technical issues</p>
-  <img src="/images/jitsi-screenshot-2018.jpg" width="100%" class="load-hangout load-hangout-screenshot"/>
-  <button class="margin-top-1 margin-bottom-1 click-text btn btn-cb2 load-hangout">{{_ "click_to_load_hangout"}}</button>
-
-</div>
-
-<div class="alert alert-info margin-top-1 margin-bottom-1">
-  <p class="align-center">If you're seeing a blank screen, <a href="https://meet.jit.si/{{room}}#config.startWithVideoMuted=true" target="_blank" id="joinHere">join here.</a> (New to Jitsi? Check out <a href="/images/codebuddies_jitsi_diagram.jpg" target="_blank">this diagram</a>.)</p>
-</div>
-  <div id="hangout-container"></div>
-  {{#if numParticipants}}
-  <div class="margin-top-1 margin-bottom-1">
-    <p class="align-center well">{{_ "hangout_participant" count=numParticipants }}</p>
-  </div>
-  {{/if}}
-  <span id="post_id" class="hide">{{_id}}</span>
-
-
-</template>
