--- client/templates/hangout/hangout-item.html
@@ -1,81 +1,0 @@
-<template name="hangoutItem">
-  <div class="hangout-item">
-    <div class="row">
-
-      <div class="col-md-12">
-        <label class="name">
-        {{#if isHangoutInProgress start end}}
-          &nbsp; &nbsp;<i class="fas fa-dot-circle fa text-danger faa-burst animated"></i> &nbsp;
-          <a href="{{pathFor 'hangout' hangoutId=_id }}" class="hangout-topic {{# if isHangoutCompleted end }}completed{{/if}}"> {{topic}} </a>
-        {{ else }}
-          <a href="{{pathFor 'hangout' hangoutId=_id }}" class="hangout-topic {{# if isHangoutCompleted end }}completed{{/if}}"> {{topic}} </a>
-        {{/if}}
-
-
-        </label>
-        {{> hangoutActionBar}}
-      </div>
-
-      <div class="col-md-4">
-
-      {{!-- Group Label --}}
-      {{#if group}}<p><a href="{{pathFor 'study group' studyGroupSlug=group.slug studyGroupId=group.id}}"> <span class="label label-default pull-right">{{group.title}}</span> </a>  </p>{{/if}}
-
-        {{#if equals visibility false}}
-          <span class="label label-pill label-danger">DELETED</span>
-        {{else}}
-
-          {{# unless isHangoutCompleted end }}
-
-            {{#if isHangoutInProgress start end}}
-
-              {{#if isAttending users}}
-                  <a class="btn btn-cb2 btn-grey btn-success" href="{{pathFor 'hangout' hangoutId=_id }}">{{_ "launch"}}</a>
-              {{else}}
-                {{# unless isHangoutCompleted end}}
-                  <button class="btn btn-cb2 join-hangout {{#unless currentUser}} continue-popup {{/unless}}">{{_ "join_in_session"}}</button>
-                {{/unless}}
-              {{/if}}
-
-            {{else}}
-
-              {{#if isAttending users}}
-                <p class="upcoming">{{upcomingTime start}}</p>
-                <a id="leave-hangout" class="btn btn-cb2 btn-leave">{{_ "leave"}}</a>
-              {{else}}
-
-                {{# unless isHangoutCompleted end}}
-                  <!-- <button id="join-hangout" class="btn btn-cb2">{{_ "rsvp"}}</button> -->
-                  <button class="btn btn-cb2 {{#if currentUser}} join-hangout {{else}} continue-popup {{/if}}">{{_ "rsvp"}}</button>
-                {{/unless}}
-
-              {{/if}}
-
-            {{/if}}
-          {{else}}
-            <p class="completed"><i class="fas fa-check" aria-hidden="true"></i> {{_ "completed"}}</p>
-          {{/unless}}
-
-        {{/if}} <!-- if equals visibility false -->
-
-    </div>
-
-    <div class="col-md-8">
-
-
-      <p class="status"><label><i class="fas {{getHangoutTypeSign type}}"></i> Started by <a href="{{pathFor 'profile' name=host.name userId=host.id}}" class="{{#unless currentUser}} continue-popup {{/unless}}">{{host.name}}</a></label></p>
-      <p class="status"><label><i class="fas fa-calendar-alt fa-lg"></i> {{getHangoutStartDateTime start}} -
-
-        {{#if isHangoutEndTimeTBA start end}}
-          To Be Announced
-        {{else}}
-          {{getHangoutEndDateTime end}}
-        {{/if}}
-          |
-         {{users.length}} Joined  {{#if views}} | <i class="fas fa-eye" aria-hidden="true" data-toggle="tooltip" title="Views" ></i> {{views}} {{/if}}</label> </p>
-      <p class="overflowtext">{{ getDescriptionTruncated description }}</p>
-    </div>
-
-  </div>
-  </div>
-</template>
