--- client/templates/hangout/hangout-learned.html
@@ -1,7 +1,0 @@
-<template name="hangoutLearned">
-	<div class="margin-bottom-1">
-	<textarea id="learned-text" rows="3" maxlength="280" class="form-control {{#unless currentUser}} continue-popup {{/unless}}" placeholder="What did you learn or accomplish?"></textarea>
-	<label id="twitterWarning"><input type="checkbox" id="chkOptInTweet"> Tweet this message automatically out on <a href="http://twitter.com/codebuddiesmeet" target="_blank"><i class="fab fa-twitter" aria-hidden="true"></i>codebuddiesmeet</a></label>
-	<span class="learnedCharactersLeft">{{learnedCharacterCount}}</span>
-	</div>
-</template>
