--- client/templates/hangout/hangout.html
@@ -1,135 +1,0 @@
-<template name="hangout">
-  <div id="post" class="container-fluid">
-    <div id="show-sidebar">
-        <a href="#">Show details</a>
-    </div>
-    <div class="row">
-
-      {{#if Template.subscriptionsReady}}
-
-        {{#with hangout}}
-
-          <!-- hangout sidebar -->
-          <div class="col-md-3 well hangout-sidebar">
-
-            <h2 class="{{# if isHangoutCompleted end }}completed{{/if}}" id="hangout-topic">
-
-              {{#if equals visibility false}}
-                <span class="label label-pill label-danger">DELETED</span>
-              {{/if}} <!-- ./if equals visibility false -->
-              <!-- close sidebar -->
-              <div id="hide-sidebar">
-                <a href="#" class="pull-right"><i class="fas fa-times-circle" aria-hidden="true"></i></a>
-              </div>
-
-              {{#if isHangoutInProgress start end}} <h5 class="in-progress"><span class="oval"></span> LIVE</h5> {{/if}}
-
-              {{topic}}
-
-              {{#if views}} <span class="view"><i class="fas fa-eye" aria-hidden="true" data-toggle="tooltip" title="Views" > </i> {{views}}</span> {{/if}}
-
-            </h2>
-
-
-            <!-- hangout group -->
-            {{#if group}}
-              <p><a href="{{pathFor 'study group' studyGroupSlug=group.slug studyGroupId=group.id}}"> <span class="label label-default">{{group.title}}</span></a></p>
-            {{/if}}
-            <!-- hangout type host -->
-            <p class="status">
-              <i class="fas {{getHangoutTypeSign type}}"></i> {{type}} hangout <sup><a id="hangout-faq-popup"><i class="fas fa-question-circle"></i></a></sup> started by
-              <a href="{{#if currentUser}} {{pathFor 'profile' name=host.name userId=host.id}} {{/if}}" class="{{#unless currentUser}} continue-popup {{/unless}}">{{host.name}}</a>
-            </p>
-
-            <!--  hangout date time -->
-            <p class="date-holder">
-              <i class="fas fa-clock" aria-hidden="true"></i>
-              <strong>{{getHangoutStartDateDay start}}</strong>
-              <br />
-
-              {{getHangoutStartTime start}} -
-              {{#if isHangoutEndTimeTBA start end}}
-                  TBA (End Manually)
-              {{else}}
-                  {{getHangoutEndTime end}}
-              {{/if}}
-
-
-            </p>
-            <!-- hangout actions buttons -->
-            {{> hangoutActionButtons }}
-
-            <!-- Participants -->
-            <h3>Participants</h3>
-            <div id="participants">
-              <a href="{{#if currentUser}} {{pathFor 'profile' name=host.name userId=host.id}} {{/if}}" class="{{#unless currentUser}} continue-popup {{/unless}}"><img class="img-circle" src="{{host.avatar}}"></a>
-
-              {{#each attendees}}
-                <a href="{{#if currentUser}} {{pathFor 'profile' name=name userId=id}} {{/if}}" class="{{#unless currentUser}} continue-popup {{/unless}}" ><img class="img-circle" src="{{avatar}}"></a>
-              {{/each}}
-            </div>
-            <!-- ./Participants -->
-
-            <!--HANGOUT ALTERNATIVES-->
-            {{> hangoutSidebarAlternatives}}
-
-            <!-- HANGOUT TIPS -->
-            {{> hangoutSidebarTips}}
-
-          </div>
-          <!-- ./hangout sidebar ends -->
-
-          <!-- hangout main container-->
-          <div class="col-md-9 hangout-body">
-            {{# unless isHangoutCompleted end}}
-              <p class="upcoming align-center margin-top-1 load-hangout">
-                  <i class="fas fa-dot-circle fa text-danger faa-burst animated"></i>{{upcomingTime start}}
-              </p>
-            {{ else }}
-              <p class="upcoming align-center margin-top-1">Sorry, this hangout has ended.</p>
-            {{/unless}}
-
-            {{#if isAttending users}}
-               {{# if externalCheckbox }}
-                 <a href="{{externalURL}}" class="btn btn-cb2 margin-top-1" rel="noopener" target="_blank">Join us on {{externalButtonText}} <i class="fas fa-external-link-alt fa-fw" aria-hidden="true"></i></a>
-               {{ else }}
-                  {{> hangoutFrame }}
-               {{/if}}
-                
-            {{ else }}
-              {{# if externalCheckbox }}
-                <a href="{{externalURL}}" class="btn btn-cb2 margin-top-1" rel="noopener" target="_blank">Join us on {{externalButtonText}} <i class="fas fa-external-link-alt fa-fw" aria-hidden="true"></i></a>
-              {{ else }}
-                  <a class="{{#if currentUser}}{{# unless isHangoutCompleted end}} join-hangout {{/unless }} {{else}} continue-popup {{/if}}"><img src="/images/logo-circle-cropped.png" width="65%" class="load-hangout-screenshot display-block centered margin-top-1"/></a>
-              {{/if}}
-            {{/if}}
-
-            {{> supportHangoutOrganizer organizerId=host.id}}
-            <h3>Goals of hangout {{> hangoutActionBar }}</h3>
-            <p id="hangout-description" class="margin-1">
-              {{{formatDescription this}}}
-            </p>
-            <!-- list of learnings for this hangout-->
-            {{> hangoutLearnings }}
-            <!-- accept learnings here-->
-            {{> hangoutLearned }}
-            <!-- comment section -->
-            <h3>Comments</h3>
-            {{> commentBox}}
-
-          </div>
-          <!-- ./hangout main container-->
-        {{else}}
-
-          {{> hangoutNotFound}}
-
-        {{/with}}
-
-      {{else}}
-
-        {{> loading}}
-
-      {{/if}}
-    </div>
-  </div>
-</template>
