--- client/templates/hangout/hangout_action_bar.html
@@ -1,65 +1,0 @@
-<template name="hangoutActionBar">
-
-  <div class="dropdown pull-right icon-spacing">
-    <button class="btn btn-default"
-            type="button"
-            id="dropdownMenu1"
-            data-toggle="dropdown"
-            aria-haspopup="true"
-            aria-expanded="true">
-      <span>
-        <i class="fas fa-ellipsis-v" aria-hidden="true">
-        </i>
-      </span>
-    </button>
-    <ul class="dropdown-menu" aria-labelledby="dropdownMenu1">
-        {{#unless hangoutOwner host.id}}
-          <li class="{{#if currentUser}} report-hangout {{else}} continue-popup {{/if}} ">
-            <a href="#">
-              Report
-            </a>
-          </li>
-        {{/unless}}
-        <li class="{{#if currentUser}} clone-hangout {{else}} continue-popup {{/if}}">
-          <a href="#">
-            Duplicate
-          </a>
-        </li>
-        {{#if hangoutOwner host.id}}
-          <li class="edit-hangout">
-            <a href="#">
-              Edit Details
-            </a>
-          </li>
-          <li class="delete-hangout">
-            <a href="#">
-              Delete
-            </a>
-          </li>
-          {{else}}
-
-          {{#if isInRole 'admin,moderator,volunteer' 'CB'}}
-            <li role="separator" class="divider"></li>
-            <li class="edit-hangout">
-              <a href="#">
-                Edit
-                <span class="label label-info">M</span>
-              </a>
-            </li>
-          {{/if}}
-          {{#if isInRole 'admin,moderator' 'CB'}}
-            <li role="separator" class="divider"></li>
-            <li class="delete-hangout">
-              <a href="#">
-                Delete
-                <span class="label label-info">M</span>
-              </a>
-            </li>
-          {{/if}}
-
-        {{/if}}
-
-    </ul>
-  </div>
-
-</template>
