--- client/templates/hangout/hangout_action_buttons.html
@@ -1,63 +1,0 @@
-<template name="hangoutActionButtons">
-  <div class="hangout-action-buttons">
-
-    {{# unless isHangoutCompleted end }}
-      <a href="{{ googleCalendarUrl this }}" rel="noopener" target="_blank" class="btn btn-block"><i class="fas fa-calendar-alt" aria-hidden="true"></i> {{_ "add_google_calendar"}}</a>
-      <a href="{{ icsDownloadLink this }}" download="hangout.ics" class="btn btn-block"><i class="fas fa-calendar-alt" aria-hidden="true"></i>{{_ "add_to_icalendar"}}</a>
-    {{/ unless }}
-
-    {{# unless isHangoutCompleted end }}
-          {{#if isAttending users}}
-            <div id="attendance">
-              <a id="leave-hangout" class="btn btn-cb2">{{_ "leave"}}</a>
-            </div>
-          {{else}}
-
-              {{# unless isHangoutCompleted end}}
-                  <button class="btn btn-cb2 {{#if currentUser}} join-hangout {{else}} continue-popup {{/if}}">{{_ "rsvp_long"}}</button>
-              {{/unless}}
-
-          {{/if}}
-
-
-    {{ else }}
-
-      <div class="completed align-center margin-top-1">
-        {{#if isAttending users}}
-            <p>Thanks for joining! This hangout has ended. Would you like to start another one?</p> <div class="btn btn-cb2 margin-top-1 {{#unless currentUser}} continue-popup  {{else}} create-hangout-popup {{/unless}}"> {{_ "create_a_hangout"}} </div>
-        {{else}}
-            <p>Sorry, this hangout has ended. Why not start your own?</p>
-            <div class="btn btn-cb2 margin-top-1 btn-schedule-hangout {{#unless currentUser}} continue-popup {{else}} create-hangout-popup {{/unless}}">
-              {{_ "create_a_hangout"}}
-            </div>
-        {{/if}}
-      </div>
-    {{/ unless }}
-
-  <div class="hangout-action-buttons-other">
-    <a href="#" class="{{#if currentUser}} clone-hangout {{else}} continue-popup {{/if}} btn btn-cb2"><i class="fas fa-clone" aria-hidden="true"></i> Duplicate</a>
-    {{#if hangoutOwner host.id}}
-      <a href="#" class="edit-hangout btn btn-cb2">Edit Details</a>
-      <a href="#" class="delete-hangout btn btn-cb2"><i class="fas fa-trash-alt" aria-hidden="true"></i> Cancel Hangout</a>
-    {{else}}
-      {{#if isInRole 'admin,moderator,volunteer' 'CB'}}
-          <a href="#" class="edit-hangout btn btn-cb2">Edit Details <span class="label label-info">M</span> </a>
-      {{/if}}
-      {{#if isInRole 'admin,moderator' 'CB'}}
-        <a href="#" class="delete-hangout btn btn-cb2 btn-danger"><i class="fas fa-trash-alt" aria-hidden="true"></i>&nbsp; Delete <span class="label label-info">M</span> </a>
-      {{/if}}
-    {{/if}}
-
-  </div>
-
-  {{#if isHangoutInProgress start end}}
-      {{#if hangoutOwner host.id}}
-          <div class="hangout-action-buttons"><div id="end-hangout" class="btn btn-cb2 btn-danger">{{_ "end_hangout"}}</div></div>
-      {{else}}
-          {{#if isInRole 'admin,moderator,volunteer' 'CB'}}
-              <div class="hangout-action-buttons"><div id="end-hangout" class="btn btn-cb2 btn-danger">{{_ 'end_hangout'}}</div></div>
-          {{/if}}
-      {{/if}}
-  {{/if}}
-</div>
-</template>
