--- client/templates/hangout/hangout_cards.html
@@ -1,42 +1,0 @@
-<template name="hangoutCards">
-
-  {{#if liveHangoutsCount }}
-    <h3 class="hangouts-title">{{_ "upcoming_events"}}</h3>
-  {{/if}}
-
-  <div id="active_hangouts">
-    {{#each hangouts 1}}
-      {{#unless isHangoutCompleted end}}
-      <div class="card">
-        {{> hangoutCard}}
-      </div>
-      {{/unless}}
-    {{/each}}
-  </div>
-
-  {{#if pastHangoutsCount }}
-    <h3 class="hangouts-title">{{_ "past_events"}}</h3>
-  {{/if}}
-
-  <div id="past_hangouts">
-    {{#each hangouts -1}}
-      {{#if isHangoutCompleted end}}
-       <div class="card">
-         {{> hangoutCard}}
-      </div>
-      {{/if}}
-    {{/each}}
-  </div>
-
-  {{#unless status}}
-    <div class="clear">
-      <button id="loadMore" class="btn btn-default btn-block">Load More</button>
-    </div>
-  {{else}}
-      <h5 class="align-center color-white">There are no more hangouts.</h5>
-  {{/unless}}
-
-  {{#unless Template.subscriptionsReady}}
-    {{> loading}}
-  {{/unless}}
-</template>
