--- client/templates/hangout/hangout_sidebar_alternatives.html
@@ -1,19 +1,0 @@
-<template name="hangoutSidebarAlternatives">
-  {{# unless isHangoutCompleted end }}
-    {{#if isHangoutInProgress start end}}
-      <h3>HANGOUT ALTERNATIVES</h3>
-      {{#if isAttending users}}
-        <p class="align-center">Want to use Google Hangouts instead?
-          {{#unless url }}
-            <div class="google-hangout display-inline">
-                <div class="g-hangout" data-render="createhangout" data-initial_apps="[{ app_id : '999356798349', start_data: {{_id}}, 'app_type' : 'LOCAL_APP' }]" data-widget_size="72"></div>
-            </div>
-          {{ else }}
-            <a class="btn btn-cb2 btn-grey btn-success google-hangout-button" href="{{url}}" target="_blank">Launch Google Hangout</a>
-          {{/unless}}
-        </p>
-      {{/if}}
-
-    {{/if}} <!-- This refers isHangoutInProgress -->
-  {{/unless}}
-</template>
