--- client/templates/hangout/hangout_sidebar_tips.html
@@ -1,14 +1,0 @@
-<template name="hangoutSidebarTips">
-  <h3>Hangout Tips</h3>
-
-  <ol>
-    <li>Remember that this is a safe space, and we help each other learn faster. Be patient, and always feel free to ask
-      questions.</li>
-    <li>You agree not to stream copyrighted or otherwise objectionable content in the hangout.</li>
-    <li class="load-hangout">When it's time, click to load the hangout. </li>
-    <li>Please keep video turned off.</li>
-    <li>Take advantage of the ability to screenshare simultaneously! It is helpful for reading the same code, pair
-      programming, and staying on task.</li>
-    <li>Remember to share what you learned.</li>
-  </ol>
-</template>
