--- client/templates/hangout/support_hangout_organizer.html
@@ -1,36 +1,0 @@
-<template name="supportHangoutOrganizer">
-    {{#with supportLinks}}
-      <div class="alert alert-warning margin-top-1 margin-bottom-1 alert-dismissible">
-        <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-
-        Did you learn a lot in this hangout?
-           <br> You can show the hangout organizer your appreciation by:
-            <ul>
-               {{# if buymeacoffee}}
-                   <li><i class="fas fa-arrow-right" aria-hidden="true"></i> <a href="https://www.buymeacoffee.com/{{buymeacoffee}}" target="_blank">buying them a coffee <i class="fas fa-coffee" aria-hidden="true"></i></a></li>
-                {{/if}}
-                {{# if patreon}}
-                    <li><i class="fas fa-arrow-right" aria-hidden="true"></i> <a href="{{patreon}}" target="_blank">supporting them on Patreon <i class="fab fa-patreon"></i></a></li>
-                {{/if}}
-                {{# if nonprofit}}
-                    <li><i class="fas fa-arrow-right" aria-hidden="true"></i> <a href="{{nonprofit}}" target="_blank">supporting a non-profit they champion <i class="fas fa-heart" aria-hidden="true"></i></a></li>
-                {{/if}}
-            </ul>
-        </div>
-    {{else}}
-      {{#if equals currentUser._id instance.data.organizerId}}
-      <div class="alert alert-warning margin-top-1 margin-bottom-1 alert-dismissible">
-        <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
-        <p>
-           <i class="fas fa-coffee" aria-hidden="true"></i> Please set up your <a href="http://buymeacoffee.com" target="_blank">Buy Me A Coffee</a> or <a href="http://patreon.com" target="_blank">Patreon</a> links so the community can tip you for being a great hangout organizer!
-           <a href="#" id='editProfileInfo'>Edit Profile</a>
-        </p>
-        <p>
-          You can also choose to feature a non-profit you'd like to support.
-        </p>
-      </div>
-      {{/if}}
-
-    {{/with}}
-
-</template>
